added team quit and join logic
